Add feature to find matching OQ curves
The OQ curve on the back is done first, but if it doesn't fit the front
then it needs to be redone for both.

This function checks if P and Q are inside the curves and if they are it
simply returns the flipped version for the front.

If they're not inside on the front then it makes a new curve for the front,
then it flips that curve for the back.