Add point along a curve (#77)
Add code for T', which is a point along a curve
This measures a certain distance along a curve from a starting point

This also included getting a curve segment from a midpoint, which could
be made a whole lot easier if information about curves and points was
kept in a history object within those classes. Then I could just remember
that T is on curve E'S between points 1005 and 1006 and not have to find
it again.